+++ R/TreeSearch_utilities.R
@@ -3,11 +3,11 @@
 #' null character, coded with state zero for every leaf in `tree`.
 #' @importFrom TreeTools MatrixToPhyDat NTip TipLabels
 #' @export
-EmptyPhyDat <- function (tree) {
+EmptyPhyDat <- function(tree) {
   mat <- matrix(0, NTip(tree), 1, dimnames = list(TipLabels(tree), NULL))
   MatrixToPhyDat(mat)
 }
 
 #' @rdname TreeSearch
 #' @export
-DoNothing <- function (...) {}
+DoNothing <- function(x = NULL) {x}