+++ tests/testthat/test-tree_length.R
@@ -261,6 +261,8 @@ test_that("CharacterLength() fails gracefully", {
   # no error:
   CharacterLength(as.phylo(1, 43, tipLabels = names(dataset)), dataset)
   
+  skip_if_not_installed("phangorn")
+  library("phangorn") # for phyDat subsetting
   expect_equal(c(53, 59, 6),
                as.numeric(table(CharacterLength(NJTree(dataset[1:4, ]),
                                                 dataset[1:4], compress = TRUE))))
@@ -273,10 +275,11 @@ test_that("Character compression works", {
   tree <- TreeTools::NJTree(dataset)
   expect_equal(length(CharacterLength(tree, dataset)), 137)
   expect_equal(length(MinimumLength(dataset)), 137)
-  expect_equal(dim(Consistency(dataset, tree)), c(137, 3))
+  expect_equal(dim(Consistency(dataset, tree, nRelabel = 0)), c(137, 4))
   expect_equal(length(CharacterLength(tree, dataset, compress = TRUE)), 118)
   expect_equal(length(MinimumLength(dataset, compress = TRUE)), 118)
-  expect_equal(dim(Consistency(dataset, tree, compress = TRUE)), c(118, 3))
+  expect_equal(dim(Consistency(dataset, tree, nRelabel = 0, compress = TRUE)),
+               c(118, 4))
 })
 
 test_that("X_MorphyLength", {